requirements/base.txt
django==3.1.2
djangorestframework==3.12.1
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.8.6
factory-boy==3.1.0
